requirement.txt
einops
torchsummary
hdf5storage
h5py
albumentations
pytorch_msssim